correction to task 7

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -13,7 +13,7 @@
 
 void puts_half(char *str)
 {
-	int length = 0;
+	long int length = 0;
 	int i;
 
 	while (str[length] != '\0')
